requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
fastapi-cache2
pyodbc
